Создание новых тестов для задания №7, попытка №3
