+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/appmanager/NavigationHelper.java
@@ -1,11 +1,6 @@
 package ru.stqa.pft.addressbook.appmanager;
 
 public class NavigationHelper extends BaseHelper {
-    private GroupHelper groupHelper;
-
-    public NavigationHelper(GroupHelper groupHelper) {
-        this.groupHelper = groupHelper;
-    }
 
     public void goToGroupPage() {
         click(find("[href='group.php']"));
