+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/tests/GroupCreationTest.java
@@ -1,6 +1,7 @@
 package ru.stqa.pft.addressbook.tests;
 
 import org.junit.jupiter.api.Test;
+import ru.stqa.pft.addressbook.appmanager.BaseHelper;
 import ru.stqa.pft.addressbook.model.GroupData;
 
 public class GroupCreationTest extends TestBase {
@@ -9,8 +10,8 @@ public class GroupCreationTest extends TestBase {
      void groupCreation() {
         app.getNavigationHelper().goToGroupPage();
         GroupData group = new GroupData("Test Group", "Test Header", "Test Footer");
-        app.getGroupHelper().createGroup(group);
-        app.verifyGroupCreated();
+        app.getCreateGroupHelper().createGroup(group);
+        BaseHelper.getTextFromMsgBox();
     }
 
 }
